fix: move server into src
This should have been done from the start, however I thought that
separating the server from the app would be better. Keeping `server`
outside `src` could lead to code duplication wrt data shared between
client and sever (e.g. regexes) or very ugly imports in the server or
the client to avoid code duplication. To fix said problems, move
`server` into `src`.
