+++ src/server/models/patient.js
@@ -1,5 +1,5 @@
 const mongoose = require('mongoose');
-const { regex } = require('../globals');
+const { regex } = require('../../globals');
 
 const p = new mongoose.Schema({
   _id: mongoose.Schema.Types.ObjectId,
